make stimulus utils module
